Add 960x960 mode to SDL 1.2 + fix LTO

+++ src/video/dc/memset.c
@@ -5,7 +5,7 @@
 
 // Set 1 byte at a time
 
-void * memset_ (void *dest, const uint8_t val, size_t len)
+void * memset_ (void *dest, const int val, size_t len)
 {
   if(!len)
   {
